% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senado-package.R
\docType{package}
\name{senado-package}
\alias{senado}
\alias{senado-package}
\title{senado: Access Data from the Brazilian Federal Senate Open Data API}
\description{
Provides a tidy interface to the Brazilian Federal Senate (Senado Federal) Open Data API. Functions return tibbles with standardized column names, automatic retry, caching, and rate limiting. Covers senators, votes, bills, committees, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SidneyBissoli/senado}
  \item Report bugs at \url{https://github.com/SidneyBissoli/senado/issues}
}

}
\author{
\strong{Maintainer}: Sidney da Silva Pereira Bissoli \email{sbissoli76@gmail.com}

}
\keyword{internal}
